Reviewing and refactoring Jest tests

+++ test/unit/processing/returns/imps/all-imps-acknowledgements-received.test.js
@@ -8,7 +8,6 @@ describe('all IMPS acknowledgements received', () => {
 
   beforeEach(async () => {
     transaction = await db.sequelize.transaction()
-
     await db.impsBatchNumber.bulkCreate([
       { batchNumber: '1', invoiceNumber: 'S123456789A123456V001', frn: 1234567890 },
       { batchNumber: '1', invoiceNumber: 'S123456789B123456V001', frn: 1234567891 },
@@ -24,38 +23,24 @@ describe('all IMPS acknowledgements received', () => {
     await db.impsAcknowledgement.destroy({ where: {}, truncate: true })
   })
 
-  test('returns true if all acknowledgements are received', async () => {
-    const acknowledgements = [
+  test.each([
+    ['all acknowledgements received', [
       { batchNumber: '1', invoiceNumber: 'S123456789A123456V001', frn: 1234567890 },
       { batchNumber: '1', invoiceNumber: 'S123456789B123456V001', frn: 1234567891 },
       { batchNumber: '1', invoiceNumber: 'S123456789C123456V001', frn: 1234567892 },
       { batchNumber: '1', invoiceNumber: 'S123456789D123456V001', frn: 1234567893 }
-    ]
-    const result = await allImpsAcknowledgementsReceived(acknowledgements, 1, transaction)
-    expect(result).toBe(true)
-  })
-
-  test('returns false if not all acknowledgements are received', async () => {
-    const acknowledgements = [
+    ], true],
+    ['partial acknowledgements received', [
       { batchNumber: '1', invoiceNumber: 'S123456789A123456V001', frn: 1234567890 },
       { batchNumber: '1', invoiceNumber: 'S123456789B123456V001', frn: 1234567891 }
-    ]
-    const result = await allImpsAcknowledgementsReceived(acknowledgements, 1, transaction)
-    expect(result).toBe(false)
-  })
-
-  test('returns false if no acknowledgements are received', async () => {
-    const acknowledgements = []
-    const result = await allImpsAcknowledgementsReceived(acknowledgements, 1, transaction)
-    expect(result).toBe(false)
-  })
-
-  test('returns false if invoicesForSequence.length is 0', async () => {
-    await db.impsBatchNumber.destroy({ where: { batchNumber: '1' }, transaction })
-    const acknowledgements = [
+    ], false],
+    ['no acknowledgements received', [], false],
+    ['no invoices in sequence', [
       { batchNumber: '1', invoiceNumber: 'S123456789A123456V001', frn: 1234567890 }
-    ]
+    ], false, true]
+  ])('%s', async (_, acknowledgements, expected, clearBatch = false) => {
+    if (clearBatch) await db.impsBatchNumber.destroy({ where: { batchNumber: '1' }, transaction })
     const result = await allImpsAcknowledgementsReceived(acknowledgements, 1, transaction)
-    expect(result).toBe(false)
+    expect(result).toBe(expected)
   })
 })
